Replaced `double` with `decimal`
This commit replaces all `double` propertiers with `decimal` if the represent a financial value.

+++ src/LexOfficeSharpApi/Model/Payment/LexPayment.cs
@@ -13,7 +13,7 @@ namespace AndreasReitberger.API.LexOffice
         DateTimeOffset postingDate;
 
         [ObservableProperty]
-        double amount;
+        decimal amount;
 
         [ObservableProperty]
         string currency = string.Empty;
